requirements.txt
mss
keyboard
pyautogui
pillow